#8974 - Group aggregated reporting data (mSERS) by jurisdiction and epi week - changes after review
