+++ sormas-api/src/main/java/de/symeda/sormas/api/report/AggregateReportGroupingLevel.java
@@ -27,6 +27,7 @@ public enum AggregateReportGroupingLevel {
 		}
 	}
 
+	@Override
 	public String toString() {
 		return I18nProperties.getEnumCaption(this);
 	}
