+++ sormas-api/src/main/resources/enum.properties
@@ -95,6 +95,12 @@ AgeGroup.AGE_75_79 = 75--79
 AgeGroup.AGE_80_84 = 80--84
 AgeGroup.AGE_80_PLUS = 80+
 
+#AggregatedReportGroupingLevel
+AggregateReportGroupingLevel.REGION = Region
+AggregateReportGroupingLevel.DISTRICT = District
+AggregateReportGroupingLevel.HEALTH_FACILITY = Facility
+AggregateReportGroupingLevel.POINT_OF_ENTRY = Point of entry
+
 # AnimalCondition
 AnimalCondition.ALIVE = Alive
 AnimalCondition.DEAD = Dead
